refactor(common): common 내부에 entity, application 패키지 추가
