+++ src/main/java/kitchenpos/order/domain/TableGroup.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.common.BaseIdEntity;
+import kitchenpos.common.entity.BaseIdEntity;
 import kitchenpos.order.application.TableGroupValidationException;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
