+++ src/test/java/kitchenpos/common/entity/QuantityTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.common;
+package kitchenpos.common.entity;
 
+import kitchenpos.common.application.ValidationException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
