+++ src/test/java/kitchenpos/order/application/TableGroupServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.application;
 
-import kitchenpos.common.NotFoundException;
+import kitchenpos.common.application.NotFoundException;
 import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.domain.TableGroupRepository;
 import kitchenpos.order.dto.OrderTableRequest_Create;
