+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.order.domain;
 
+import kitchenpos.common.entity.Quantity;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.Product;
 import kitchenpos.order.application.OrderValidationException;
-import kitchenpos.common.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
